Fixed some bugs, added printf's and added callback functions

+++ OnionPayloadData.cpp
@@ -35,7 +35,7 @@ OnionPayloadData::~OnionPayloadData() {
     
 }
 
-int OnionPayloadData::getRawLength() {
+uint16_t OnionPayloadData::getRawLength() {
     return this->rawLength;
 }
